gst-avsynctest-rs: extract reusable A/V sync analysis helpers
Promote the recovery-side measurements shared by the standalone generator
tests and the MXL av_sync round-trip -- audio pip detection, video bar
centroid, and CEA-708 caption decoding -- into a public `analyze` module.
Both test suites now call it instead of carrying private copies, and any
other transport under test can assert A/V and caption alignment on recovered
samples the same way.

Behaviour is unchanged: the functions are moved verbatim, only made public
and pointed at the crate's own `captions` constants.
